Change default target type
When using ECS, you have to select ip. And since it is thought that ECS (or EKS) will become the default in the near future, we decided that it is preferable to set ip to the default value.

Based on https://docs.aws.amazon.com/AmazonECS/latest/developerguide/service-load-balancing.html#load-balancing-concepts